Avoid re-downloading profile pic with --latest-stamps
Using --latest-stamps allows one to delete the downloaded files from a
previous session and on the next run download only new items. The
exception is the profile picture, which is always downloaded again if
the file was deleted.

I've added a control to avoid downloading the profile pic if it has
been downloaded before. Since there is no timestamp on the profile
pic, the filename is saved. If on the next run the profile pic has not
changed, it is not downloaded again. If it has changed, and it
downloaded normally.